object client - improve stream behavior and perf
removed the redundant readaheads which were practically the same as the
stream’s buffering - instead we let the stream’s big buffer (32 mb) do
the ordered readahead and for concurrency we now read the next 8
consecutive ranges in parallel.

also forced closing the reader on req/res close so it won’t have
remains of canceled requests.

finally disable the http forced truncate of ranged requests. this
apparently is preventing the browser from doing it’s client side
buffering correctly.