two RPC bugs
1. integers were limited to 32bit because of bitwise | usage for
truncating numbers to integers, fixed to use parseInt which is a bit
slower but works for big numbers as well. this was needed to support
time integers which are bigger than 32bit integers.

2. fix stuck http response when in browser got status 0 (connection
refused), and in that case the response object does not emit any
events, so just avoid registering response events in that case and
immediately throw to reject promise chains.