disable mongoose autoIndex in schemas
instead we call explicitly on web server.
this is to allow deferring on production, and also for tests that call
this right after dropping the db.
in newer versions of mongoose it is possible to set this in a single
setting, but for now we gotta disable on every schema.