fix #568 fix dedup to avoid mixing different encryptions on the same chunk
the fix is that if the dup chunk is unavailable to allocate new chunk.
the changes to the driver and the object_coding are not relevant and
just remains from previous attempt and didn’t remove to keep the option
to use more than single object_codings in the future.