Fix n2n wrong connection url used on signal
added the source url to the signal and redirect calls,
so that the receiver can assign the new connection url to the source,
since this url is used as the key for the rpc connection map.

also refactored the redirector code to separate the add/remove agent functions
to be used from several paths and keep the global registry map and the connection set consistent.