TEMP! delete object in case of error while uploading
Since we don’t have recover from write first block, we fail easily. We
must solve it soon. For now, we will have this patch.

Fuse in such case, can’t handle the response we send and we are stuck
with zero size object with no parts.

deleting the object allows better recovery in this case.