fix node._id used as string when saved in blocks by map builder, which fixes read_node_mappings which was always returning empty.
this issue was introduced once node_allocator was changed to use node_monitor instead of directly from mongo, so any older database will not have this issue, and no db upgrade is needed.

also fixed in other queries and in general to avoid using node._id as string when not intended.