FE - Fix Most Issues in Firefox
NooBaa FE now fully supports Firefox.

The only issue left is that firefox do not allow the developer to style
it’s scrollbars via css so the scrollbars may look different.

There is a good chance that this fix will also enable a Usable FE in
in Microsoft Edge but i didn’t checked this yet.

#388