Fixing issue #1505
There was a loophole in the cluster master variable.
When we raised the debug level it actually did raise it, but because
the server was not the master (even when there is no cluster you need
to get the master value), we did not update in the db.
Which meant that the frontend did not know that the system in debug
mode.