Added tests to test build chunks and infrastructure changes in test
Object_io will check in verification mode that the server sees the same
digest for all blocks of the same object part.
mongoldb defaults now includes resetting error reports in agents.
Added missing test cases to test_build_chunks.
Changed the verification of object health to include more tests:
Will make sure that the file was not tampered with, that the cloud node
is ready, that the object is mapped to cloud node if it is supposed to
be, and that the actual blocks exist on the cloud pool. These tests
currently require more nodes because they are untrusted after tests
that corrupt blocks on them.
Moved the function that checks blocks exist on cloud to a new test util
file.