Added connection usage to account schema
Added connection usage to account schema to allow front end to display
and decide whether or not to allow user to delete the connection from
the user.
Also changed pool and bucket schema to hold connection id instead of
connection as it is not really being used and not guaranteed to be
defined.