fixed rebuild issues
- fixed load_parts_objects_for_chunks which is called by map_builder,
to also find deleted objects and parts, so they will not cause node
migration to get stuck
- in map_utils.get_chunk_status, separated allocations of special
replicas to a different allocations array which is handled as an
optional allocation that doesn’t block rebuilds\migrations. chunk is
considered good even if it doesn’t have all of the extra allocations
stored.