Disappearing
Forcing recreation of the <use> element (inside svg-icon> each time the
icon name (href value) changes. This is done due to the fact that
chrome 58 broke behavior that cause changes to xlink:href to be ignored
which manifested itself as empty icons (issue #2994)

This need to be coupled with a cache control for icons.svg (currently
only out of dev mode) so the icons that change value will not flicker.