fix #3176 - support copy from file to same same name (used for set metadata attributes)
the fix is to move the target file deletion to complete_multipart_upload.
this was actually a historic misunderstanding that we implemented it in create_multipart_upload,
and also AWS only deletes the target file on complete.
so we added the upload_started field to the unique index to allow multiple objects with same key to exist.

mongo upgrade - drop old objectmds unique key index, and move mark completed to run as last script