setup swap on upgrade
The logic added to our upgrade will check for the status of the swap,
and if no swap will allocate 8GB swapfile and use it.

In ESX/OVA we have LVM with swap partition allocated by default.
This fix is for Azure VMs (and other cloud VMs) where we don't have swap partition as our image doesn't use LVM at all.
The reason LVM is not used on Azure is that it is not recommended to be used on OS disk.
See https://docs.microsoft.com/en-us/azure/virtual-machines/linux/configure-lvm.