Change placement logic of create_func when no pools are given as params
This change was done because the user may not understand that the new
function is using the resource (it was not specified by him) and may
tamper with the resource (remove nodes or delete the resource) also it
fixes the case where the account has no default resources (in case of
an admin account with no s3 access).

By allowing the create function to set a funciton with no pools
attached we fix both of these problems.

Also rewrite create_func as an async function

Move some of the func prop calculations to the top

This was done so the usage of pools_by_name will be near the the
extraction site (form req.system)