Move most of APIClient interface to its promise chain
This change will decrease the size of an APiClient object to an object with a 2 fields an rpc reference and the options for the client.
The interface objects (which contain the method declaration) will be generated on the fly on each access and will consist of a single field named `client` that point back to the client that created them.
The methods themselves are now implemented on the prototype of each API object.

A side effect of this change is that we cannot pass an API method directly as a callback because it loses its context. A workaround is to pass an arrow function that preserves the context. See the changes to node.n2n_singal as an example.