jenkins: initial noobaa-core CentOS CI job converted from Travis CI
The noobaa-core Jenkins job runs two tests in parallel on a bare metal
machine reserved through 'cico' in the CentOS CI. The jobs are based on
the configuration in `.travis.yml`.

Steps done in the `noobaa-core.groovy` Jenkins Pipeline:
  - checkout the repository (for scripts in the .jenkins/ directory)
  - reserve a bare metal machine in the CI
  - copy to and run the `prepare.sh` script on the bare metal machine
  - start the two jobs
    - unit test
    - build
  - return the bare metal machine to the CI pool

The `.jenkins/jobs/noobaa-core.yaml` is an example of how the Jenkins
Pipeline can be imported with Jenkins Jobs Builder. It is currently
unused.

This configuration of jobs is based on the Ceph-CSI projects' ci/centos
branch.

Signed-off-by: Niels de Vos <ndevos@redhat.com>