jenkins: install an use Docker CE instead of Podman
minikube does not work with Podman yet. At the moment Podman and Docker
CE can not get installed on the same machine, so we'll use Docker for
now.

In the CentOS CI Jenkins instance, the bare-metal machines fail to build
the tester image due to a conflict with the default IP-address of the
container. This seems to be a problem while building only, not during
running the tests. Pass `--network=host` to `docker build` when the
Jenkins jobs run it.

Fixes: #6199
Signed-off-by: Niels de Vos <ndevos@redhat.com>
(cherry picked from commit a2d232de17b0125e3bc60aefd6f535491b65b9a1)