Prebuild lib files and include it in the repository
Refactor the lib build process into an offline task. Upon lib dependencies change the developer should run this task on his dev environment and include the results (the src/lib folder) in the commit that is pushed to the repository.

This removes any dependencies on bower in our upstream and downstream automated build processes. automated build processes.

Signed-off-by: nb-ohad <mitrani.ohad@gmail.com>