Installing nasm via wget instead of PowerTools repo
Installing nasm via wget instead of PowerTools repo. Currently, we are pacing some issues with centos PowerTools repo and as a resault, the tests are failing on build. This PR will switch to installing nasm via wget to avoid those issues. Once PowerTools repo is stable again we can switch back to it.

Signed-off-by: liranmauda <liran.mauda@gmail.com>