Disable Centos Jenkins jobs
Disable Centos Jenkins jobs

Jenkins has an issue with pulling and GH Action is doing what we currently need, so we are disabling the Centos Jenkins jobs.
We might want to add to the Centos Jenkins jobs new tests (if GH Actions can’t handle the load) so we are keeping the code.

Signed-off-by: liranmauda <liran.mauda@gmail.com>