fixing issue of NOENT when uploading
fix issue where upload failes with "The specified key does not exist" when in the same time delete happens that deletes the target directory (as part of being the last object there). In that case we will retry the rename and not just fail.
Partly fix to #6766

Signed-off-by: jackyalbo <jacky.albo@gmail.com>
(cherry picked from commit 0162d411d50fe1fd0e62174d38fbaececdaadc18)