Add support to content_encoding
We didn't support content_encoding and didn't return it to the user in case the user supplied it when uploading the file.
Now, if provided, we will save the encoding in the DB and return it to the user on head/getObject.

Signed-off-by: jackyalbo <jacky.albo@gmail.com>
(cherry picked from commit bef76aba442fc9d2054786d557a0c611c8714be1)