nooba/nooba-core: Nooba bucket Logging configuration
This change provides the basic schema and bucket API
which could be used for configuring logging of a bucket.

Target Bucket:
A different bucket should be provided to store logs of
a bucket. These logs can be used for "Log based replication"

Log Prefix:
A prefix is provided which will be added to log records to
differentiate logs of a bucket.

Signed-off-by: Ashish Pandey <aspandey@redhat.com>