Fix paths in RPM
Some files were previously copied to the builder container instead of
storing them in the tarball as you should.

This commit fixes that and adds the files to the tarball. It also
unifies the way files like these are handled in the spec file and
creates symbolic links for all such files.

Signed-off-by: Boris Ranto <Boris.Ranto@ibm.com>