bucket_logging: Change bucket API's and functions as per schema
We do not store name of the bucket, on which we are setting bucket
logging, in bucket schema. There is no need to return that name
as part of get_bucket_logging and store it as part of
put_bucket_logging

Also, made changed to return a correct value in case of bucket
logging is configured for a bucket.

Signed-off-by: Ashish Pandey <aspandey@redhat.com>