Bucket_logging: Time bound bucket logs delivery
Log rotation, segregation and upload of log objects for different
buckets happens based on size and time. If the size of logs are less
than the configured value, it might take time to upload the log objects.

We need to consider the following factors behind log collection and processing.
1 - logrotation depends on size of the log file, bucket_logs.log, or time of
the day. Currently, the size is set to 100KB. logrotate wakes up every 1 minute.
That means, once the bucket_logs.log file size reaches more than 100KB, logrotates
will come up and rotate this file. If the size has not reached 100K, it still
rotates the file after every 24hr.

2 - After rotation, bucket_logs.log gets processed by logsegregate.sh which gets
executed every 5 min. This creates log objects for individual buckets. We have a
limit of size of log objects which is 50K. Once a log object reaches 50K, it will
be transferred to the log bucket.

3 - Background process, which actually copies the log object in log buckets,
scans the log path every 5 minutes. If it finds any log object, it copies it to a
log bucket.

If log is not delivered because it is not rotated in step 1 based on
size, then that can only be rotated on daily basis. We can not do
anything there.

However, we have control on step 2. If the size of the log object for a bucket
is less than 50K, it is not being uploaded even after 2-3 hours or even 24 hr.

Solution: We need to make sure that even if the size of log object for a bucket
is less than 50K, it get uploaded once the threshold of time reaches,
which is set to be 5 minutes.

Signed-off-by: Ashish Pandey <aspandey@redhat.com>