bucket logging: Change logrotate configuration from daily to hourly
Changing conf to rotate the log file every hour if it meets
minimum size criteria.

hourly
minsize 10
maxsize 50k

It means that for every hour (01:00 AM, 02:00 AM and so on) when
logrotate is called, it will see if the size of the log is atleast
10 bytes or not. If it is, it will rotate the logs.

maxsize makes sure that if the log size is 50k or more than as soon
as logrotate is executed this log will be rotated. Even if it happens
every minute or 5 minutes or so.

Signed-off-by: Ashish Pandey <aspandey@redhat.com>