Bucket Logging: Accept Log Prefix as per S3 bucket logging
We need to change the log prefix restriction and accept
it in accordance with AWS S3 guidelines.

As we do that, it is better to fetch the log prefix from
the configuration of the source bucket instead of depending
on log object names.

It's because the log prefix can have any character in it, and
it would be difficult to have a delimiter that can segregate
the log prefix from other strings in the log object name.

Signed-off-by: Ashish Pandey <aspandey@redhat.com>