Adding unsupported ceph nsfs s3 test to pending list
This is the missed commit from PR:
https://github.com/noobaa/noobaa-core/pull/8201

The PR #8062 is returning the 'DisplayName' taking it from
'bucket_info.bucket_owner where noobaa is preserving the the
actual bucket owner info. Ealier we were just returning what
user was providing and test was passing as it the expected
value was getting reflected in the response.

The test is failing now since we are returning actually info
coming from the bucket_info.bucket_owner and not from the 'owner'
field of the parent object.

testname name that is moved to pending_lists is
s3tests_boto3/functional/test_s3.py::test_bucket_list_return_data

Signed-off-by: Vinayakswami Hariharmath <vharihar@redhat.com>